Render metadata inside the tooltip
The metadata is expected to be in a JSON format describing
how to render the tag inside the table row. It has a name
field which describes the row name, a tag field describing
the tag to create and a innerHTML field describing the
text to render inside the tag, finally theres an attributes
field which sets the attributes of the tag to the key:values
inside it.

The schema is then {name: string, tag: string, innerHTML: string,
attributes: {attribute-name: attribute-value (string)}}